Add manual trigger and input control to GitHub Actions
Added a `workflow_dispatch` trigger to the GitHub Actions workflow, allowing the workflow to be manually triggered. Introduced an input parameter `run_matrix` under `workflow_dispatch` to control whether the matrix job should run. Updated the `matrix_job` to include a comment explaining that the job will only run if the `run_matrix` input is set to 'true'.